refactor(viewer): split TrialDetailPage into trial components and hook
- Keep page as route composition only
- Extract header, tabs, trajectory, file-split panels, and use-trial-detail (Refs: #31)
