+++ apps/viewer/src/components/trial/trial-header.tsx
@@ -1,0 +1,94 @@
+import { ChevronLeft, ChevronRight } from "lucide-react";
+
+import { Button } from "@/components/ui/button";
+import type { Trial } from "@/lib/api";
+
+export function TrialHeader({
+  runId,
+  taskId,
+  trial,
+  prevId,
+  nextId,
+  onSibling,
+}: {
+  runId: string;
+  taskId: string;
+  trial: Trial | null;
+  prevId: string | null;
+  nextId: string | null;
+  onSibling: (id: string | null) => void;
+}) {
+  return (
+    <div className="flex flex-wrap items-start justify-between gap-3">
+      <div className="min-w-0 flex-1">
+        <h1 className="text-2xl font-semibold tracking-tight text-ink font-mono truncate">
+          {runId}
+        </h1>
+        {/* Same mute/body style: task · framework · docker · upstream(if any) */}
+        <p className="text-sm text-mute mt-1 flex flex-wrap items-center gap-x-1.5 gap-y-0.5">
+          <span>
+            task <span className="text-body font-medium">{taskId}</span>
+          </span>
+          {trial?.framework ? (
+            <>
+              <span className="text-mute select-none" aria-hidden>
+                ·
+              </span>
+              <span className="text-body font-medium font-mono text-[13px]">
+                {trial.framework}
+              </span>
+            </>
+          ) : null}
+          {trial?.docker ? (
+            <>
+              <span className="text-mute select-none" aria-hidden>
+                ·
+              </span>
+              <span className="text-body font-medium font-mono text-[13px]">
+                {trial.docker}
+              </span>
+            </>
+          ) : null}
+          {trial?.upstream_url ? (
+            <>
+              <span className="text-mute select-none" aria-hidden>
+                ·
+              </span>
+              <a
+                href={trial.upstream_url}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="text-link hover:text-link-deep font-mono text-[13px] truncate max-w-[min(48ch,100%)]"
+                title={trial.upstream_name || trial.upstream_url}
+              >
+                {trial.upstream_url}
+              </a>
+            </>
+          ) : null}
+        </p>
+      </div>
+      <div className="flex items-center gap-1 shrink-0">
+        <Button
+          type="button"
+          variant="outline"
+          size="icon"
+          disabled={!prevId}
+          onClick={() => onSibling(prevId)}
+          aria-label="Previous trial"
+        >
+          <ChevronLeft className="h-4 w-4" />
+        </Button>
+        <Button
+          type="button"
+          variant="outline"
+          size="icon"
+          disabled={!nextId}
+          onClick={() => onSibling(nextId)}
+          aria-label="Next trial"
+        >
+          <ChevronRight className="h-4 w-4" />
+        </Button>
+      </div>
+    </div>
+  );
+}
