+++ apps/viewer/src/components/trial/types.ts
@@ -1,0 +1,11 @@
+import type { Trial } from "@/lib/api";
+
+export type ActorRow = NonNullable<Trial["actors"]>[number];
+
+export function actorLabel(a: ActorRow | undefined, profileId: string): string {
+  if (!a) return profileId;
+  const bits = [a.role || profileId];
+  if (a.agent) bits.push(String(a.agent));
+  if (a.model) bits.push(String(a.model));
+  return bits.join(" · ");
+}
